Fix Chars functions
Adds / Deletions :
 - Remove ft_chrswitch (useless)

Changes :
 - Use “chars” instead of “chrs”
 - Use “unsigned int” instead of “size_t” for parameters that are not a
size but numbers of chars.
 - Remove useless NULL or end-of-string tests
 - Adapt to the norm
 - Optimize iterations (decrement when possible)

+++ libft/basics/chars/output/ft_putspace.c
@@ -6,14 +6,14 @@
 /*   By: ncoden <ncoden@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/18 11:48:38 by ncoden            #+#    #+#             */
-/*   Updated: 2015/08/27 23:23:51 by ncoden           ###   ########.fr       */
+/*   Updated: 2015/08/31 16:15:45 by ncoden           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 #include "libft/basics/chars.h"
 
-void		ft_putspace(size_t len)
+void			ft_putspace(unsigned int len)
 {
 	while (len > 0)
 	{
